Test for serializing full json metadata
